+++ Navigation/Assets.xcassets/AppIcon.appiconset/Contents.json
