+++ Navigation/Assets.xcassets/cat1.imageset/Contents.json
