+++ Navigation/Assets.xcassets/logo.imageset/Contents.json
