+++ Navigation/Assets.xcassets/superdog.imageset/superdog.jpg
