+++ Navigation/PostSecond.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct PostSecond {
+    let author: String
+    let description: String
+    let image: String
+    let likes: Int
+    let views: Int
+}
+
+let posts: [PostSecond] = [
+    PostSecond(author: "Котолюб", description: "Милый котик ждет своей очереди к ветеринару.Милый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринаруМилый котик ждет своей очереди к ветеринару", image: "cat1", likes: 240, views: 312),
+    PostSecond(author: "Собакалюб", description: "Собака породы овчарка", image: "dog1", likes: 766, views: 893)
+]
