+++ Navigation/PostTableViewCell.swift
@@ -1,0 +1,116 @@
+import UIKit
+
+final class PostTableViewCell: UITableViewCell {
+    
+    private let postImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFit
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.backgroundColor = .black
+        return imageView
+    }()
+    
+    private let authorLabel: UILabel = {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 20, weight: .bold)
+        label.numberOfLines = 2
+        label.textColor = .black
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+   
+    private let descriptionLabel: UILabel = {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 14)
+        label.textColor = .systemGray
+        label.numberOfLines = 0
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    private let likesLabel: UILabel = {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 16)
+        label.textColor = .black
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    private let viewsLabel: UILabel = {
+        let label = UILabel()
+        label.font = .systemFont(ofSize: 16)
+        label.textColor = .black
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    private let bottomStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        return stackView
+    }()
+    
+    private let contentStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.spacing = 12
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        return stackView
+    }()
+    
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupViews()
+        setupConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupViews() {
+        contentStackView.addArrangedSubview(authorLabel)
+        contentStackView.addArrangedSubview(postImageView)
+        
+        contentView.addSubview(contentStackView)
+        contentView.addSubview(descriptionLabel)
+
+        bottomStackView.addArrangedSubview(likesLabel)
+        bottomStackView.addArrangedSubview(viewsLabel)
+        
+        contentView.addSubview(bottomStackView)
+    }
+    
+    private func setupConstraints() {
+        NSLayoutConstraint.activate([
+            contentStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            contentStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            contentStackView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 16),
+            
+            descriptionLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            descriptionLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            descriptionLabel.topAnchor.constraint(equalTo: contentStackView.bottomAnchor, constant: 16),
+            
+            bottomStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            bottomStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            bottomStackView.topAnchor.constraint(equalTo: descriptionLabel.bottomAnchor, constant: 16),
+            bottomStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -16),
+            
+            postImageView.heightAnchor.constraint(equalToConstant: 300)
+        ])
+    }
+
+    
+    func configure(with post: PostSecond) {
+        authorLabel.text = post.author
+        authorLabel.textColor = .black
+        
+        descriptionLabel.text = post.description
+        
+        postImageView.image = UIImage(named: post.image)
+        
+        likesLabel.text = "Likes: \(post.likes)"
+        viewsLabel.text = "Views: \(post.views)"
+    }
+}
