+++ Navigation/ProfileHeaderView.swift
@@ -1,7 +1,7 @@
 import UIKit
 
 class ProfileHeaderView: UIView {
-   private let avatarImageView: UIImageView = {
+    private let avatarImageView: UIImageView = {
         let img = UIImageView()
         img.layer.borderWidth = 3
         img.layer.borderColor = UIColor.white.cgColor
@@ -10,8 +10,8 @@ class ProfileHeaderView: UIView {
         img.layer.cornerRadius = 50
         img.clipsToBounds = true
         img.image = UIImage(named: "superdog")
-     return img
-   }()
+        return img
+    }()
     
     private let fullNameLabel: UILabel = {
         let label = UILabel()
@@ -31,63 +31,69 @@ class ProfileHeaderView: UIView {
         return label
     }()
     
-     private let statusTextField: UITextField = {
+    private let statusTextField: UITextField = {
         let textField = UITextField()
-        textField.placeholder = "Enter new status"
+        textField.placeholder = "Set your status..."
+        textField.borderStyle = .none
         textField.font = .systemFont(ofSize: 14, weight: .regular)
         textField.textColor = .black
         textField.backgroundColor = .white
         textField.layer.cornerRadius = 12
+        textField.layer.borderColor = UIColor.black.cgColor
+        textField.layer.borderWidth = 1.0
         textField.translatesAutoresizingMaskIntoConstraints = false
+        textField.clipsToBounds = true
         return textField
     }()
+
     
     private let setStatusButton: UIButton = {
         let btn = UIButton()
-        btn.setTitle("Set Status", for: .normal)
-        btn.backgroundColor = .blue
+        btn.setTitle("Set status", for: .normal)
+        btn.backgroundColor = .blueSecondary
         btn.setTitleColor(.white, for: .normal)
-        btn.layer.cornerRadius = 4
+        btn.layer.cornerRadius = 10
         btn.layer.shadowOffset = CGSize(width: 4, height: 4)
         btn.layer.shadowRadius = 4
         btn.layer.shadowColor = UIColor.black.cgColor
         btn.layer.shadowOpacity = 0.7
         btn.translatesAutoresizingMaskIntoConstraints = false
         return btn
-   }()
+    }()
     
     override init(frame: CGRect) {
-      super.init(frame: frame)
+        super.init(frame: frame)
         setupConstraints()
         setStatusButton.addTarget(self, action: #selector(onButtonTap), for: .touchUpInside)
+        backgroundColor = .white
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
+    
     private func setupConstraints() {
         addSubview(avatarImageView)
         addSubview(fullNameLabel)
         addSubview(statusLabel)
         addSubview(statusTextField)
         addSubview(setStatusButton)
-
-       NSLayoutConstraint.activate([
-           avatarImageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
-           avatarImageView.topAnchor.constraint(equalTo: topAnchor),
-           avatarImageView.widthAnchor.constraint(equalToConstant: 100),
-           avatarImageView.heightAnchor.constraint(equalToConstant: 100),
+        
+        NSLayoutConstraint.activate([
+            avatarImageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+            avatarImageView.topAnchor.constraint(equalTo: topAnchor),
+            avatarImageView.widthAnchor.constraint(equalToConstant: 100),
+            avatarImageView.heightAnchor.constraint(equalToConstant: 100),
             
             fullNameLabel.leadingAnchor.constraint(equalTo: avatarImageView.trailingAnchor, constant: 16),
-            fullNameLabel.topAnchor.constraint(equalTo: topAnchor, constant: 10),
+            fullNameLabel.topAnchor.constraint(equalTo: topAnchor),
             fullNameLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
             
             statusLabel.leadingAnchor.constraint(equalTo: avatarImageView.trailingAnchor, constant: 16),
-            statusLabel.topAnchor.constraint(equalTo: fullNameLabel.bottomAnchor, constant: 10),
+            statusLabel.bottomAnchor.constraint(equalTo: statusTextField.topAnchor, constant: -10),
             statusLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
             
-            statusTextField.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+            statusTextField.leadingAnchor.constraint(equalTo: avatarImageView.trailingAnchor, constant: 16),
             statusTextField.topAnchor.constraint(equalTo: avatarImageView.bottomAnchor, constant: 16),
             statusTextField.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
             statusTextField.heightAnchor.constraint(equalToConstant: 40),
@@ -102,6 +108,8 @@ class ProfileHeaderView: UIView {
     @objc private func onButtonTap() {
         statusLabel.text = statusTextField.text
     }
- }
-  
+}
 
+#Preview {
+    ProfileHeaderView()
+}
