+++ Navigation/SceneDelegate.swift
@@ -1,7 +1,6 @@
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-    
     var window: UIWindow?
     
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
@@ -55,10 +54,5 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         func sceneDidEnterBackground(_ scene: UIScene) {
             
         }
-        
-        
     }
-    
-    
 }
-
