--- Navigation/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Navigation
-//
-//  Created by Виктория Вайнштейн on 25.10.2024.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
